move validations to validate folder
