+++ packages/cli/src/util/validate/validateMethod.ts
